video_predictor/requirements.txt
boto3
Flask
python-dotenv
opencv-python
pandas
numpy
keras
tensorflow
